move test code to sub directory
