+++ src/ema/mod.rs
@@ -1,12 +1,6 @@
 mod math;
 
 #[cfg(test)]
-pub(crate) mod high_precision;
-#[cfg(test)]
-mod invariants;
-#[cfg(test)]
-mod test_data;
-#[cfg(test)]
 mod tests;
 
 pub use math::*;
