+++ src/ema/tests/high_precision.rs
